requirements.txt
pyudev
psutil
ExifRead
python-magic